+++ internal/aws/imagebuilder/distribution_configuration_resource_gen.go
@@ -22,6 +22,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/hashicorp/terraform-provider-awscc/internal/generic"
 	"github.com/hashicorp/terraform-provider-awscc/internal/registry"
+	fwvalidators "github.com/hashicorp/terraform-provider-awscc/internal/validators"
 )
 
 func init() {
@@ -274,6 +275,37 @@ func distributionConfigurationResource(ctx context.Context) (resource.Resource, 
 		//	      "Region": {
 		//	        "description": "region",
 		//	        "type": "string"
+		//	      },
+		//	      "SsmParameterConfigurations": {
+		//	        "description": "The SSM parameter configurations to use for AMI distribution.",
+		//	        "insertionOrder": true,
+		//	        "items": {
+		//	          "additionalProperties": false,
+		//	          "description": "The SSM parameter configuration for AMI distribution.",
+		//	          "properties": {
+		//	            "AmiAccountId": {
+		//	              "description": "The account ID for the AMI to update the parameter with.",
+		//	              "type": "string"
+		//	            },
+		//	            "DataType": {
+		//	              "description": "The data type of the SSM parameter.",
+		//	              "enum": [
+		//	                "text",
+		//	                "aws:ec2:image"
+		//	              ],
+		//	              "type": "string"
+		//	            },
+		//	            "ParameterName": {
+		//	              "description": "The name of the SSM parameter.",
+		//	              "type": "string"
+		//	            }
+		//	          },
+		//	          "required": [
+		//	            "ParameterName"
+		//	          ],
+		//	          "type": "object"
+		//	        },
+		//	        "type": "array"
 		//	      }
 		//	    },
 		//	    "required": [
@@ -617,6 +649,55 @@ func distributionConfigurationResource(ctx context.Context) (resource.Resource, 
 						Description: "region",
 						Required:    true,
 					}, /*END ATTRIBUTE*/
+					// Property: SsmParameterConfigurations
+					"ssm_parameter_configurations": schema.ListNestedAttribute{ /*START ATTRIBUTE*/
+						NestedObject: schema.NestedAttributeObject{ /*START NESTED OBJECT*/
+							Attributes: map[string]schema.Attribute{ /*START SCHEMA*/
+								// Property: AmiAccountId
+								"ami_account_id": schema.StringAttribute{ /*START ATTRIBUTE*/
+									Description: "The account ID for the AMI to update the parameter with.",
+									Optional:    true,
+									Computed:    true,
+									PlanModifiers: []planmodifier.String{ /*START PLAN MODIFIERS*/
+										stringplanmodifier.UseStateForUnknown(),
+									}, /*END PLAN MODIFIERS*/
+								}, /*END ATTRIBUTE*/
+								// Property: DataType
+								"data_type": schema.StringAttribute{ /*START ATTRIBUTE*/
+									Description: "The data type of the SSM parameter.",
+									Optional:    true,
+									Computed:    true,
+									Validators: []validator.String{ /*START VALIDATORS*/
+										stringvalidator.OneOf(
+											"text",
+											"aws:ec2:image",
+										),
+									}, /*END VALIDATORS*/
+									PlanModifiers: []planmodifier.String{ /*START PLAN MODIFIERS*/
+										stringplanmodifier.UseStateForUnknown(),
+									}, /*END PLAN MODIFIERS*/
+								}, /*END ATTRIBUTE*/
+								// Property: ParameterName
+								"parameter_name": schema.StringAttribute{ /*START ATTRIBUTE*/
+									Description: "The name of the SSM parameter.",
+									Optional:    true,
+									Computed:    true,
+									Validators: []validator.String{ /*START VALIDATORS*/
+										fwvalidators.NotNullString(),
+									}, /*END VALIDATORS*/
+									PlanModifiers: []planmodifier.String{ /*START PLAN MODIFIERS*/
+										stringplanmodifier.UseStateForUnknown(),
+									}, /*END PLAN MODIFIERS*/
+								}, /*END ATTRIBUTE*/
+							}, /*END SCHEMA*/
+						}, /*END NESTED OBJECT*/
+						Description: "The SSM parameter configurations to use for AMI distribution.",
+						Optional:    true,
+						Computed:    true,
+						PlanModifiers: []planmodifier.List{ /*START PLAN MODIFIERS*/
+							listplanmodifier.UseStateForUnknown(),
+						}, /*END PLAN MODIFIERS*/
+					}, /*END ATTRIBUTE*/
 				}, /*END SCHEMA*/
 			}, /*END NESTED OBJECT*/
 			Description: "The distributions of the distribution configuration.",
@@ -682,11 +763,13 @@ func distributionConfigurationResource(ctx context.Context) (resource.Resource, 
 	opts = opts.WithTerraformSchema(schema)
 	opts = opts.WithAttributeNameMap(map[string]string{
 		"account_id":                           "AccountId",
+		"ami_account_id":                       "AmiAccountId",
 		"ami_distribution_configuration":       "AmiDistributionConfiguration",
 		"ami_tags":                             "AmiTags",
 		"arn":                                  "Arn",
 		"container_distribution_configuration": "ContainerDistributionConfiguration",
 		"container_tags":                       "ContainerTags",
+		"data_type":                            "DataType",
 		"description":                          "Description",
 		"distributions":                        "Distributions",
 		"enabled":                              "Enabled",
@@ -703,11 +786,13 @@ func distributionConfigurationResource(ctx context.Context) (resource.Resource, 
 		"name":                                 "Name",
 		"organization_arns":                    "OrganizationArns",
 		"organizational_unit_arns":             "OrganizationalUnitArns",
+		"parameter_name":                       "ParameterName",
 		"region":                               "Region",
 		"repository_name":                      "RepositoryName",
 		"service":                              "Service",
 		"set_default_version":                  "SetDefaultVersion",
 		"snapshot_configuration":               "SnapshotConfiguration",
+		"ssm_parameter_configurations":         "SsmParameterConfigurations",
 		"tags":                                 "Tags",
 		"target_account_ids":                   "TargetAccountIds",
 		"target_repository":                    "TargetRepository",